+++ app/src/main/java/com/julia/bachelor/SalgTemplate.java
@@ -1,0 +1,29 @@
+package com.julia.bachelor;
+
+import java.io.Serializable;
+
+interface SalgTemplate {
+    long get_ID();
+
+    void set_ID(long _ID);
+
+    String getKunde();
+
+    void setKunde(String kunde);
+
+    String getDato();
+
+    void setDato(String dato);
+
+    String getVarer();
+
+    void setVarer(String varer);
+
+    int getBelop();
+
+    void setBelop(int belop);
+
+    String getBetaling();
+
+    void setBetaling(String betaling);
+}
